added title and cc scenes, fixed alignment issues with dialogue heads, changed placement of sprites/text
